Increase the Docker daemon timeout
It takes more than 10s (potentially up to 30, but let's start with 20)
for Docker to start up the first time while it sets up the devicemapper
stuff. The Mesos instances are generally okay with that, but the Docker
registries try to pull down an image immediately after installing
Docker, so they end up timing out.